jate temos tmu

+++ set_files.sh
@@ -1,11 +1,13 @@
-
-
 PATHS=$(pwd)
 
-
 echo $PATHS
-rm ~/.zshrc
-ln -s $PATHS/terminal/.zshrc ~/.zshrc
+
+if [ "$1" -eq 1 ]; then
+	rm ~/.zshrc
+	ln -s $PATHS/terminal/.zshrc ~/.zshrc
+else
+	echo "nao modou .zshrc"
+fi
 
 rm ~/.config/terminator/config
 ln -s $PATHS/terminitor/config ~/.config/terminator/config
@@ -13,3 +15,11 @@ ln -s $PATHS/terminitor/config ~/.config/terminator/config
 rm -rf ~/.config/nvim
 ln -s $PATHS/nvim/ ~/.config/nvim
 
+rm -rf ~/.tmux
+ln -s $PATHS/tmux/.tmux ~/.tmux
+
+rm -rf ~/.tmux.conf
+ln -s $PATHS/tmux/.tmux.conf ~/.tumx.conf
+
+rm -rf ~/.tmux.conf.local
+ln -s $PATHS/tmux/.tmux.conf.local ~/.tmux.conf.local
